Alter default log levels to exclude INFO, set mastering data as default test data (it's a larger dataset)
