+++ Lib/fontbakery/frontend_spa_report/src/Settings.ts
@@ -29,7 +29,7 @@ export enum IkonType {
  * TODO: Use log levels from Font Bakery??
  *  Would have to output them to the HTML as a JS var.
  */
-export const FontBakeryLogLevels: FontBakeryStatus[] = ['INFO', 'WARN', 'ERROR', 'FAIL'];
+export const FontBakeryLogLevels: FontBakeryStatus[] = ['WARN', 'ERROR', 'FAIL'];
 
 export const CheckStatusEmoji: Record<FontBakeryStatus, string> = {
     ERROR: '💥',
