+++ Lib/fontbakery/frontend_spa_report/src/components/FontBakeryReport.vue
@@ -32,6 +32,7 @@ const filteredSections = computed(() => {
                 )
             "
         />
+        <hr />
     </template>
 </template>
